// frontend/src/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        DogRead: {
            /** Name */
            name: string;
            /** Breed */
            breed: string;
            /** Age */
            age: number;
            /** Sex */
            sex: string;
            /** Size */
            size: string;
            /** Description */
            description: string;
            /** Photo Url */
            photo_url: string;
            /**
             * Status
             * @default available
             */
            status: string;
            /** Id */
            id: number;
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /**
             * Updated At
             * Format: date-time
             */
            updated_at: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}